SERVER-1625 Remove unnecessary changes, making mongo-nonx86 more like the original.

+++ db/btree.h
@@ -36,9 +36,9 @@ namespace mongo {
         little<unsigned short> _kdo;
 
         short keyDataOfs() const {
-            return _kdo;
+            return (short) _kdo;
         }
-        
+
         void setKeyDataOfs(short s) {
             _kdo = s;
             assert(s>=0);
